Phase 2 prep: move pure-data parser layer to flutter_scene_importer
Phase 2 (build-time integration of the runtime importer) requires
flutter_scene_importer's build hook to use the parser code, but
flutter_scene_importer can't depend on flutter_scene (that direction
exists already, would be circular).

Move the four no-engine-dep files from flutter_scene to
flutter_scene_importer:

  packages/flutter_scene/lib/src/runtime_importer/glb.dart
  packages/flutter_scene/lib/src/runtime_importer/gltf_types.dart
  packages/flutter_scene/lib/src/runtime_importer/gltf_parser.dart
  packages/flutter_scene/lib/src/runtime_importer/accessor.dart
    →
  packages/flutter_scene_importer/lib/src/gltf/glb.dart
  packages/flutter_scene_importer/lib/src/gltf/types.dart
  packages/flutter_scene_importer/lib/src/gltf/parser.dart
  packages/flutter_scene_importer/lib/src/gltf/accessor.dart

Expose them publicly via packages/flutter_scene_importer/lib/gltf.dart.

The engine-type builders (geometry_builder, material_builder,
texture_builder, skin_builder, animation_builder, runtime_importer)
stay in flutter_scene since they need engine types (Node, Mesh,
Material, gpu.Texture). They now import the parser layer via
package:flutter_scene_importer/gltf.dart.

No behavior change. All 19 tests pass.
