+++ packages/flutter_scene_importer/lib/src/gltf/glb.dart
